Experimental free space function
The experimental function kind of works. Currently, the angle where the goal should be placed is correct, but, the distance (x and y pos) must be scaled so that the goal is not into the wall. The problem with this is that at some points, this scalled value is in a wall even if the goal set is good. This occurs when the scale value is at 10%, 25%, 50% 75%, etc.. The current idea to fix is this, is to 1. recognize the scaled points are in a wall 2. move the goal from that point, towards the middle of the corridor.