Experimental free space function 2
This new experimental free space function works by finding the average theta. This is done by summing each vector in the laserscan from -90 to 0 with its positive counterpart on each pass. For each pair, the theta relative to the x vector of the kart is stored and they are all averaged at the end. The current issue revolves around knowing when to weight the thetas corresp to the lazerscans closer to the side more than those closer to the front of the kart. It was found that too much weighting on the side will cause the kart to turn 90 to one side and eventually cause the goal to be into the wall (biggest failure mode). The other thing that can happen though is when having to take a sharp turn: in this case the front thetas should be weighted much less than the side ones in order to avoid a goal into the wall.