goal setting exp method 2 working!
The fix needed was to recognize when the front of the kart was to close to the wall. At that point, reducing the magnitude of the x and y portions of the vector to a reasonable amount allowed the kart to slow down (closer goals to the robot) and use goals farther out from the robot when in straights.