Remove dependency install script (use rosdep).
After multiple necessary additions to the dependency install script as well as the discovery of many missing dependencies in the CMakelist.txt and package.xml files of some of our sub-packages, I decided that it would be best to use rosdep in order to install all the package dependencies. This way, we rely on ROS to figure out what packages each person needs to install and it also ensures that our CMakelists and package.xml files are up to date with the packages that we depend on throughout our code.