Remove the global tf params (will delete later once design is confirmed to work).
For clarity sake, it is better to define these in each of the parameter files that need to know these frames. This way, when someone just looks at one of these param files, they will know what the frames are. This coule be problematic if we wanted to change the names of the tf frames but we should always be following rep 105 (https://www.ros.org/reps/rep-0105.html) and thus this should be constant.