Discovered issue of hector_mapping not being able to see te static transform from base_link to velodyne frame.
In addition to adding the --clock flag when running the rosbag, hect_mapping needs the parameter simtime to be set to true when using bagged LiDAR data. See the comment placed at the top of hector_mapping.launch and learn more in the ROS Clock API docs http://wiki.ros.org/Clock.