Failed attempt to try to remove/ignore stageros' frames.
I tried to use tf_remap to change the names of the tf frames so that our nodes can publish the tfs we have like in the kart launch files. But this does not work because there is no good way to remove tfs from stageros. See a pretty good explination here: https://answers.ros.org/question/175013/removingdeletingignoring-odom-tf-data/. Like option 1 in that post, I am also going to try to directly modify stageros to force it to not publish the base_footprint and odom frames.