Fix serial pkt and control pkt memcpy issue
This fixes the issue with memcpy call on control packet by changin control packet v_speed and v_angle from int to char in amp_serial_jetson.h. Previously this caused an issue where serial packet looked in msg[0] and msg[1] bytes to get the values of speed and angle, but the control packet was an 8 byte struct of 2 integers which caused msg[0] and msg[1] to be the values of the first int in the struct.